updated Grid
