--- client/src/components/Grid/Col.js
@@ -1,16 +1,0 @@
-import React from "react";
-
-// This Col component lets us size bootstrap columns with less syntax
-// e.g. <Col size="md-12"> instead of <div className="col-md-12">
-export function Col({ size, children }) {
-  return (
-    <div
-      className={size
-        .split(" ")
-        .map(size => "col-" + size)
-        .join(" ")}
-    >
-      {children}
-    </div>
-  );
-}
